Allow skipping token validity checks
Provide a new option to tell the `googleAuthR` library not to perform its
standard authorization validations. Users of `googleAuthR` can enable
this option with:

```r
options(googleAuthR.skip_token_checks = TRUE)
```

This feature is desirable in certain contexts, like testing. For
example, `googleCloudStorageR` could be configured to connect to a GCS
emulator that doesn't require or implement any authorization. In that
situation `googleCloudStorageR` could use this new option to opt out of
unnecessary checks in `googleAuthR`.